Making proper use of size_t
In a lot of cases, int was being used instead of size_to represent
sizes. That led to some warnings about inconsistencies between int
and size_t usage. While this cl doesn't solve all warnings, it tries
to use size_t and int where it should be appropriate to use them.

+++ src/OpenGL/compiler/Intermediate.cpp
@@ -1053,8 +1053,8 @@ bool CompareStruct(const TType& leftNodeType, ConstantUnion* rightUnionArray, Co
     int index = 0;
 
     for (size_t j = 0; j < structSize; j++) {
-        int size = fields[j]->type()->getObjectSize();
-        for (int i = 0; i < size; i++) {
+        size_t size = fields[j]->type()->getObjectSize();
+        for(size_t i = 0; i < size; i++) {
             if (fields[j]->type()->getBasicType() == EbtStruct) {
                 if (!CompareStructure(*(fields[j]->type()), &rightUnionArray[index], &leftUnionArray[index]))
                     return false;
@@ -1078,7 +1078,7 @@ bool CompareStructure(const TType& leftNodeType, ConstantUnion* rightUnionArray,
         int arraySize = leftNodeType.getArraySize();
 
         for (int i = 0; i < arraySize; ++i) {
-            int offset = typeWithoutArrayness.getObjectSize() * i;
+            size_t offset = typeWithoutArrayness.getObjectSize() * i;
             if (!CompareStruct(typeWithoutArrayness, &rightUnionArray[offset], &leftUnionArray[offset]))
                 return false;
         }
@@ -1253,7 +1253,7 @@ ConstantUnion* CreateInverse(TIntermConstantUnion* node, ConstantUnion* unionArr
 TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNode, TInfoSink& infoSink)
 {
     ConstantUnion *unionArray = getUnionArrayPointer();
-    int objectSize = getType().getObjectSize();
+    size_t objectSize = getType().getObjectSize();
 
     if (constantNode) {  // binary operations
         TIntermConstantUnion *node = constantNode->getAsConstantUnion();
@@ -1678,12 +1678,11 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
 
 TIntermTyped* TIntermediate::promoteConstantUnion(TBasicType promoteTo, TIntermConstantUnion* node)
 {
-    int size = node->getType().getObjectSize();
+    size_t size = node->getType().getObjectSize();
 
     ConstantUnion *leftUnionArray = new ConstantUnion[size];
 
-    for (int i=0; i < size; i++) {
-
+    for(size_t i = 0; i < size; i++) {
         switch (promoteTo) {
             case EbtFloat:
                 switch (node->getType().getBasicType()) {
